requirements.txt
Django==5.1.1
Pillow==11.1.0
